Extracts LoginPopup into separate component, which can be unhidden using the UserContext
