+++ src/Components/Login/Login.css
@@ -1,52 +1,3 @@
-#login-popup {
-    position: fixed;
-    height: 100%;
-    width: 100%;
-    margin: 0;
-    top: 0;
-    left: 0;
-    z-index: 1;
-    background-color: #000c;
-}
-
-#login-popup > form {
-    opacity: 1;
-    border: 1px solid white;
-    width: 18rem;
-    height: 15rem;
-    background-color: #000;
-    border-radius: 0.5rem;
-}
-
-#signin-fields {
-    margin-top: 2rem;
-    height: 80%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    align-items: center;
-    text-align: center;
-    border: none;
-}
-
-#login-popup input {
-    width: 50%;
-    margin-left: 0.5rem;
-}
-
-#login-popup p {
-    height: 1rem;
-    color: #f44;
-}
-
-#login-close {
-    position: relative;
-    float: right;
-    width: 1.5rem;
-    height: 1.5rem;
-    padding: 0;
-}
-
 #user-display {
     display: flex;
     align-items: center;
