+++ src/Contexts/UserContext.jsx
@@ -5,10 +5,18 @@ export const UserContext = createContext();
 export const UserProvider = ({ children }) => {
     const [user, setUser] = useState({});
     const [isLoggedIn, setIsLoggedIn] = useState(false);
+    const [showLoginPopup, setShowLoginPopup] = useState(false);
 
     return (
         <UserContext.Provider
-            value={{ user, setUser, isLoggedIn, setIsLoggedIn }}
+            value={{
+                user,
+                setUser,
+                isLoggedIn,
+                setIsLoggedIn,
+                showLoginPopup,
+                setShowLoginPopup,
+            }}
         >
             {children}
         </UserContext.Provider>
